requirements.txt
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.1.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==5.0.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.3
dacite==1.8.1
dataclasses-json==0.5.8
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.103.0
fastjsonschema==2.17.1
filelock==3.12.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.6.0
galai==1.1.6
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.0.1
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
-e git+https://github.com/jrichardson97/kg-summarizer.git@2cd46847c6cbac12d89f9a5ab92970c617c67bcf#egg=kg_summarizer
langchain==0.0.218
langchainplus-sdk==0.0.17
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
neo4j==5.10.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.17.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parallelformers==1.2.7
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==1.10.9
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
reasoner-pydantic==4.1.4
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
stack-data==0.6.2
starlette==0.27.0
streamlit==1.26.0
sympy==1.12
tenacity==8.2.2
terminado==0.17.1
tiktoken==0.4.0
tinycss2==1.1.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.25.1
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.23.2
validators==0.21.2
watchdog==3.0.0
watchfiles==0.20.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
widgetsnbextension==4.0.7
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.0
